Update from the end of my day of coding.
I managed to fix up the script to deal with variations in the results tables and how they are named. A bunch of IF statements sorted that out. I'm very close to being able to send R off to webscrape a whole year's worth of race result tables and to put them into .csv files with a unique 'race.ID'.

On the database front, I've mastered the following:
a. Connecting to the MySQL server
b. Creating a database 'ProCycling'
c. Creating calendar tables 'calendar_cn_2009' etc in the database
d. Executing basic queries and putting the results into a dataframe