Database Day!
Not as much progress as I had hoped. Finding more problems with unique non UTF-8 characters.

I hvae managed to sort out the table name and column naming conventions to allow the race calendars to be successfully written to the database. I'm still having problems with assignment of the PRIMARY KEY.