After doing the Giro daily update, today I worked on further sorting out the race result tables.
The cool bit I worked out today is how to split a combined column (in this case "Rider (Country) Team" into separate columns using the 'separate' function from Hadley's 'tidyr' package. It took a bit to understand the syntax and identify the correct regex expression for matching special characters (in this case the parentheses).

Next steps:

- Combine the script I developed today to separate columns with the other new script I've developed recently to use rvest to download race result tables, correctly assign headers and assign columns for pts/time and race_id.